Rework EndpointProvider
