--- Sources/Alchemy/Alchemy+Papyrus/HTTPMethod+NIO.swift
@@ -1,8 +1,0 @@
-import NIOHTTP1
-import Papyrus
-
-extension HTTPReqMethod {
-    var nio: HTTPMethod {
-        .init(rawValue: self.rawValue)
-    }
-}
