+++ Sources/Papyrus/Empty.swift
@@ -1,1 +1,1 @@
-public struct Empty: RequestCodable { public init() {} }
+public struct Empty: EndpointRequest { public init() {} }
