+++ Sources/Papyrus/HTTPRequestMethod.swift
@@ -1,11 +1,13 @@
-public enum HTTPReqMethod: String {
-    case CONNECT = "CONNECT"
-    case DELETE = "DELETE"
-    case GET = "GET"
-    case HEAD = "HEAD"
-    case OPTIONS = "OPTIONS"
-    case PATCH = "PATCH"
-    case POST = "POST"
-    case PUT = "PUT"
-    case TRACE = "TRACE"
+public struct EndpointMethod: Equatable {
+    public static let delete = EndpointMethod("DELETE")
+    public static let get = EndpointMethod("GET")
+    public static let patch = EndpointMethod("PATCH")
+    public static let post = EndpointMethod("POST")
+    public static let put = EndpointMethod("PUT")
+    
+    public let rawValue: String
+    
+    public init(_ rawValue: String) {
+        self.rawValue = rawValue
+    }
 }
