+++ Sources/Papyrus/PropertyWrappers/Endpoints/GET.swift
@@ -1,0 +1,19 @@
+@propertyWrapper
+public class GET<Req: EndpointRequest, Res: Codable> {
+    public var wrappedValue: Endpoint<Req, Res>
+    
+    public init(_ basePath: String) {
+        self.wrappedValue = Endpoint<Req, Res>(method: .get, path: basePath)
+    }
+    
+    public static subscript<EnclosingSelf: EndpointGroup>(
+        _enclosingInstance object: EnclosingSelf,
+        wrapped wrappedKeyPath: ReferenceWritableKeyPath<EnclosingSelf, Endpoint<Req, Res>>,
+        storage storageKeyPath: ReferenceWritableKeyPath<EnclosingSelf, GET<Req, Res>>
+    ) -> Endpoint<Req, Res> {
+        get { object[keyPath: storageKeyPath].wrappedValue.with(baseURL: object.baseURL) }
+        // This setter is needed so that the propert wrapper will have a `WritableKeyPath` for using
+        // this subscript.
+        set { fatalError("Endpoints should not be set.") }
+    }
+}
