--- Sources/Papyrus/PropertyWrappers/Methods/DELETE.swift
@@ -1,8 +1,0 @@
-@propertyWrapper
-public class DELETE<Req: RequestAllowed, Res: Codable> {
-    public var wrappedValue: Endpoint<Req, Res>
-
-    public init(_ basePath: String) {
-        self.wrappedValue = Endpoint<Req, Res>(method: .DELETE, basePath: basePath)
-    }
-}
