--- Sources/Papyrus/PropertyWrappers/Methods/PUT.swift
@@ -1,8 +1,0 @@
-@propertyWrapper
-public class PUT<Req: RequestAllowed, Res: Codable> {
-    public var wrappedValue: Endpoint<Req, Res>
-
-    public init(_ basePath: String) {
-        self.wrappedValue = Endpoint<Req, Res>(method: .PUT, basePath: basePath)
-    }
-}
